refactor: move TTL controller to where it is used
