+++ pkg/vulnerabilityreport/controller.go
@@ -11,8 +11,8 @@ import (
 	"github.com/aquasecurity/trivy-operator/pkg/apis/aquasecurity/v1alpha1"
 	"github.com/aquasecurity/trivy-operator/pkg/exposedsecretreport"
 	"github.com/aquasecurity/trivy-operator/pkg/kube"
-	"github.com/aquasecurity/trivy-operator/pkg/operator/controller"
 	"github.com/aquasecurity/trivy-operator/pkg/operator/etc"
+	"github.com/aquasecurity/trivy-operator/pkg/operator/jobs"
 	"github.com/aquasecurity/trivy-operator/pkg/trivyoperator"
 	"github.com/go-logr/logr"
 	appsv1 "k8s.io/api/apps/v1"
@@ -37,7 +37,7 @@ type WorkloadController struct {
 	etc.Config
 	client.Client
 	kube.ObjectResolver
-	controller.LimitChecker
+	jobs.LimitChecker
 	kube.LogsReader
 	kube.SecretsReader
 	Plugin
