+++ pkg/vulnerabilityreport/ttl_report.go
@@ -1,4 +1,4 @@
-package controller
+package vulnerabilityreport
 
 import (
 	"context"
